Testing Atomic variables from concurrency package
Here, AtomicInteger count, is used to avoid threads messing up with count variable. 

'Count++' is not an atomic operation. On machine level code we approximate the operations as,
temp=count
temp=temp+1
count=temp

if threads perform temp+1 and count=temp operations simultaneously then one of the temp+1 is not reflected in the count variable as neither operations waited for the other to complete the increment and take the incremented value in temp.

Atomic variable is used to avoid this situation where the operation doesn't happen in an atomic way as said above.