Synchronization using blocks or rather locks
Without synchronization this block of code can throw ArrayIndexOutOfBound exception and each ArrayList may not have 2k(intended) elements in it.

If we synchronize the entire stageOne() and stageTwo() methods then we lose the efficiency of running the program using threads. So we create Objects lock1 and lock2 and make threads obtain these locks in order to execute the block covered by it.